relocated welcome page to next js directory
