+++ src/app/welcome/page.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../../components/header/Header.js';
-import Footer from '../../components/footer/Footer.js';
-import Content from '../../app/components/Content.js';
+import Header from '../components/header/Header.js';
+import Footer from '../components/footer/Footer.js';
+import Content from '../components/Content.js';
 
 function Welcome() {
   return (<div>
